We decoupled axios.get requests for both MovieSearch Form and Rental Library, we moved them from VideoStore to their respective components
