+++ src/components/Customer.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+// import PropTypes from 'prop-types';
 
 class Customer extends Component {
   // <a href='#'>Customer List</a>
